In-process, file-free CDM
Stop doing file I/O for the CDM check. It seems slower somehow, but
that's likely because the compression algorithms being used are
stronger and because timing now forces a `doall`--previously, we may
have been timing "time taken to construct the lazy sequence" or
something.

Also change the interface to `diss` to take a `{:uuid,:trace}` map for
each sequence.